GELSVDOEBRISNDV
Elvis Aaron Presley (January 8, 1935 – August 16, 1977) was an American singer and actor. Referred to as the "King of Rock and Roll", he is regarded as one of the most significant cultural figures of the 20th century. Presley's sexually provocative performance style, combined with a mix of influences across color lines during a transformative era in race relations, brought both great success and initial controversy.